Added Structure to the legend and legend items classes
