+++ Test/MetaDrawSettingsAndViewsTest.cs
@@ -1,5 +1,6 @@
 ﻿using EngineLayer;
 using GuiFunctions;
+using GuiFunctions.ViewModels.Legends;
 using IO.MzML;
 using NUnit.Framework;
 using OxyPlot;
